Create Unique_Paths.py
A robot is located at the top-left corner of a m x n grid (marked 'Start' in the diagram below).

The robot can only move either down or right at any point in time. The robot is trying to reach the bottom-right corner of the grid (marked 'Finish' in the diagram below).

Now consider if some obstacles are added to the grids. How many unique paths would there be?

We could do this the DP way. The base cases are:
1. Robot reaching destination
2. Robot hitting an obstacle or going out of the maze / grid / board

The idea is to start from square 1 and find the answer to the subproblem how many path are there if the robot is moved one step down or to the right.